fix(clientes): ajusta imports e props do ClientesPage
