+++ MODULOS/CLIENTES.md
@@ -1,0 +1,18 @@
+# Módulo de Clientes
+
+Campos principais:
+- id
+- name
+- email
+- phone
+- document
+- address
+- city
+- state
+- postal_code
+- notes
+- is_active
+- created_at
+- updated_at
+
+Fluxo básico: listagem, criação, edição e exclusão de clientes com filtros avançados e exportação CSV.
